requirements.txt
pillow
opencv-python
pytesseract
pyautogui
requests
numpy